+++ NEWS.md
@@ -1,6 +1,7 @@
-# secretbase 0.3.0.9003 (development)
+# secretbase 0.3.0.9004 (development)
 
-* Adds SipHash-1-3 pseudorandom function (PRF).
+* Adds HMAC generation to `sha256()`.
+* Adds SipHash-1-3 pseudo-random function (PRF) as a fast, cryptographically-strong keyed hash.
 
 # secretbase 0.3.0.1